% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmdx_decode.R
\name{decode_all_sample}
\alias{decode_all_sample}
\title{Decode the output of the LLM.}
\usage{
decode_all_sample(response)
}
\arguments{
\item{response}{the character vector of the LLM response as
output of `chattr::ch_submit_job()`.}
}
\value{
a data.frame made of one line per response, entities of the taxonomy as columns.
}
\description{
Decode the output of the LLM.
}
