% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmdx_decode.R
\name{remove_layout}
\alias{remove_layout}
\alias{decode_json_result}
\alias{decode_yaml_result}
\title{Remove layout information from each entity}
\usage{
remove_layout(df)

decode_json_result(response)

decode_yaml_result(response)
}
\arguments{
\item{df}{the dataframe of parsed response}

\item{response}{the character vector of the LLM response as
output of `chattr::ch_submit_job()`.}
}
\value{
the same dataframe with layout information removed in the values

a data.frame made of one line per response, entities of the taxonomy as columns.

a data.frame made of one line per response, entities of the taxonomy as columns.
}
\description{
Remove layout information from each entity

Decode the output of the LLM with a strict json parser.

Decode the output of the LLM with a YAML parser.
}
