% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmdx_prompt.R
\name{lmdx_prompt}
\alias{lmdx_prompt}
\title{Turn a pdf document into LMDX prompt}
\usage{
lmdx_prompt(document, taxonomy, segment = "word", chunk = "page", max_lines)
}
\arguments{
\item{document}{a pdf document}

\item{taxonomy}{an entity taxonomy to extract from the document.}

\item{segment}{used to pass parameter to the underlying `lmdx_document()` function.}

\item{chunk}{either "page" or "sequence", the chunking of the text. If `page`, the default,
the original page chunk is used, with the risk of out of sequence tokens if the page 
contains a lot of words. if `sequence`, the text is chunked into lines to fit in the LLM
max-sequence length.}

\item{max_lines}{length of the chunk in lines in case of chunk = "sequence".}
}
\value{
the prompt to pass to the LLM
}
\description{
Turn a pdf document into LMDX prompt
}
