% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmdx_decode.R
\name{majority_vote}
\alias{majority_vote}
\title{Perform majority vote on each and every entity of the taxonomy}
\usage{
majority_vote(response_df)
}
\arguments{
\item{response_df}{a dataframe with one line per response, entities of the taxonomy as columns.}
}
\value{
a single line dataframe with the same columns as response_df aggregating
 the most frequent value of each column.
}
\description{
Perform majority vote on each and every entity of the taxonomy
}
