% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_2_lmdx.R
\name{pivot_longer_to_segment}
\alias{pivot_longer_to_segment}
\title{Title}
\usage{
pivot_longer_to_segment(pdf_data, segment = "word")
}
\arguments{
\item{pdf_data}{}

\item{segment}{}
}
\description{
Title
}
