% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmdx_document.R
\name{pivot_to_short_segment}
\alias{pivot_to_short_segment}
\title{Chunk a pdf_data long format into a list of wide format per segment.}
\usage{
pivot_to_short_segment(pdf_data_df, segment, max_lines = 100)
}
\arguments{
\item{segment}{either "word" , "font", or "line", the text segment to assemble}

\item{max_lines}{the text segment max chunk size in lines}

\item{pdf_data}{a data frame output of `pdftools::pdf_data()`}
}
\value{
the list of small text formatted in wide format of no more than 100 lines.
}
\description{
Chunk a pdf_data long format into a list of wide format per segment.
}
