% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmdx_decode.R
\name{remove_layout}
\alias{remove_layout}
\title{Remove layout information from each entity}
\usage{
remove_layout(df)
}
\arguments{
\item{df}{the dataframe of parsed response}
}
\value{
the same dataframe with layout information removed in the values
}
\description{
Remove layout information from each entity
}
