% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmdx_document.R, R/lmdx_prompt.R
\name{lmdx_prompt}
\alias{lmdx_prompt}
\title{Turn a pdf document into LMDX prompt}
\usage{
lmdx_prompt(document, taxonomy, ...)

lmdx_prompt(document, taxonomy, ...)
}
\arguments{
\item{document}{a pdf document}

\item{taxonomy}{an entity taxonomy to extract from the document.}

\item{...}{used to pass parameter to the underlying `lmdx_document()` function.}

\item{entities}{an entity taxonomy to extract from the document}
}
\value{
the prompt to pass to the LLM

the prompt to pass to the LLM
}
\description{
Turn a pdf document into LMDX prompt

Turn a pdf document into LMDX prompt
}
