% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmdx_document.R
\name{pivot_longer_to_segment}
\alias{pivot_longer_to_segment}
\title{Turn a pdf_data long format into wide format per segment}
\usage{
pivot_longer_to_segment(pdf_data, segment)
}
\arguments{
\item{pdf_data}{a data frame output of `pdftools::pdf_data()`}

\item{segment}{either "word" , "font", or "line", the text segment to assemble}
}
\value{
the text segmented in wide format, trailed with the segment layout values in the format "x|y".
}
\description{
Turn a pdf_data long format into wide format per segment
}
