% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmdx_document.R
\name{pivot_to_short_segment}
\alias{pivot_to_short_segment}
\title{Chunk a pdf_data long format into a list of wide format per segment.}
\usage{
pivot_to_short_segment(pdf_data_df, segment, max_lines)
}
\arguments{
\item{pdf_data_df}{a data frame of the flatten output of  `pdftools::pdf_data()`}

\item{segment}{layout information granularity. either "word", "font" or "line", the text
segment to assemble for each layout information.}

\item{max_lines}{the text segment max chunk size in lines}
}
\value{
the list of text formatted in wide format chunk-ed into no more than `max_lines`.
}
\description{
Chunk a pdf_data long format into a list of wide format per segment.
}
